chore: fix typo
