+++ scripts/release.ts
@@ -6,7 +6,7 @@ import {checkChangelogBeforePush} from './shared/check-changelog-before-push';
 import {gitCommitAndPush} from './shared/git-commit-and-push';
 import {makeReleaseBranch} from './shared/make-release-branch';
 import {TuiReleaseMode} from './shared/release-mode';
-import {runStandardVersion} from './shared/run-standard-verion';
+import {runStandardVersion} from './shared/run-standard-version';
 
 const ci = hasFlag(`--ci-mode`);
 const mode = getValueByFlag<TuiReleaseMode>(`--release-as`, `minor`);
